Upd. Code. Refactoring code base to improve performance.
